+++ src/utils/timeParser.ts
@@ -1,19 +1,28 @@
 // 7 * 86400 * 1000
 const limit: number = 604800000;
 
-export default (time: string): false | number => {
-    if(!/[0-9]+(?=s|m|h|d)/g.test(time))
-        return false;
+export default class Time {
+    public raw: string;
+    public ms: number;
+    public invalid: boolean;
 
-    const parsed: string[] = time.match(/[0-9]+|(s|m|h|d)/g);
-    let parsedTime: number = parseInt(parsed[0]);
+    constructor(raw: string) {
+        this.raw = raw;
 
-    switch(parsed[1]) {
-        case 's': parsedTime = parsedTime * 1000; break;
-        case 'm': parsedTime = parsedTime * 60000; break;
-        case 'h': parsedTime = parsedTime * 3600000; break;
-        case 'd': parsedTime = parsedTime * 86400000; break;
+        if(!/[0-9]+(?=s|m|h|d)/g.test(raw))
+            this.invalid = true;
+        else {
+            let parsed: string[] = raw.match(/[0-9]+|(s|m|h|d)/g);
+            let time: number = parseInt(parsed[0]);
+
+            switch(parsed[1]) {
+                case 's': this.ms = time * 1000; break;
+                case 'm': this.ms = time * 60000; break;
+                case 'h': this.ms = time * 3600000; break;
+                case 'd': this.ms = time * 86400000; break;
+            }
+
+            if(this.ms > limit && limit != -1) this.invalid = true;
+        }
     }
-
-    return (parsedTime > limit && limit != -1 || parsedTime < 10000) ? false : parsedTime;
 }