+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 # 2.5 - (something)
 ## 2.5.1
-- Updated `permissions` command
+- Fixed `permissions` command
 - Added `greeting` command
+- Optimization
+- *oof* Updated README version from `2.0.2` to `2.5.1`
+- Updated the time parser
 
 ## 2.5.0 (btw, this is probably the biggest update ever)
 - Fixed `rank, ranking, math` commands